Refactor Slack Sample App with new Header
